Add caching in aleph + move some functions into utilities.py
